requirements.txt
torch==1.13.1
tqdm==4.65.0
numpy==1.24.3
transformers@git+https://github.com/huggingface/transformers.git@cae78c46
Pillow==9.5.0
decord==0.6.0
gradio==3.23.0
markdown2==2.4.8
einops==0.6.1
requests==2.30.0
sentencepiece==0.1.99
protobuf==4.23.2
accelerate==0.20.3
tokenizers>=0.13.3
pydantic==1.10.7
pycocotools
pycocoevalcap
torchvision
boto3
pyyaml
punctuator==0.9.6
ffmpeg_python==0.2.0
pandas==1.1.1
ffmpeg==1.4
pipelines
scikit_learn==0.24.2
CherryPy==18.6.0
h5py==3.8.0
matplotlib==3.5.3
decord==0.6.0
opencv-python==4.7.0.72